refactor: decompose @blecsd/game index.ts into logical modules
